+++ lib/ui/page/home/tab/menu/view.dart
@@ -22,6 +22,7 @@ import '/routes.dart';
 import '/themes.dart';
 import '/ui/page/home/page/chat/widget/back_button.dart';
 import '/ui/page/home/page/my_profile/controller.dart';
+import '/ui/page/home/tab/menu/status/view.dart';
 import '/ui/page/home/widget/app_bar.dart';
 import '/ui/page/home/widget/avatar.dart';
 import '/ui/widget/widget_button.dart';
@@ -70,9 +71,7 @@ class MenuTabView extends StatelessWidget {
                 const SizedBox(width: 10),
                 Flexible(
                   child: WidgetButton(
-                    onPressed: () {
-                      // TODO: Display status changing modal.
-                    },
+                    onPressed: () => StatusView.show(context),
                     child: DefaultTextStyle.merge(
                       maxLines: 1,
                       overflow: TextOverflow.ellipsis,
@@ -87,26 +86,19 @@ class MenuTabView extends StatelessWidget {
                                   'dot'.l10n * 3,
                               style: const TextStyle(color: Colors.black),
                             ),
-                            Row(
-                              mainAxisSize: MainAxisSize.min,
-                              children: [
-                                Obx(() {
-                                  return Text(
-                                    c.myUser.value?.presence
-                                            .localizedString() ??
-                                        'dot'.l10n * 3,
-                                    style: Theme.of(context).textTheme.caption,
-                                  );
-                                }),
-                                const SizedBox(width: 2),
-                                Icon(
-                                  Icons.expand_more,
-                                  size: 18,
-                                  color:
-                                      Theme.of(context).colorScheme.secondary,
-                                ),
-                              ],
-                            ),
+                            Obx(() {
+                              return Text(
+                                c.myUser.value?.presence.localizedString() ??
+                                    'dot'.l10n * 3,
+                                style: Theme.of(context)
+                                    .textTheme
+                                    .caption
+                                    ?.copyWith(
+                                      color:
+                                          c.myUser.value?.presence.getColor(),
+                                    ),
+                              );
+                            }),
                           ],
                         );
                       }),
@@ -123,6 +115,7 @@ class MenuTabView extends StatelessWidget {
           body: Padding(
             padding: const EdgeInsets.symmetric(vertical: 5),
             child: ListView.builder(
+              key: const Key('MenuListView'),
               itemCount: ProfileTab.values.length,
               itemBuilder: (context, i) {
                 final Widget child;
@@ -216,6 +209,7 @@ class MenuTabView extends StatelessWidget {
                 switch (ProfileTab.values[i]) {
                   case ProfileTab.public:
                     child = card(
+                      key: const Key('PublicInformation'),
                       icon: Icons.person,
                       title: 'label_public_information'.l10n,
                       subtitle: 'label_public_section_hint'.l10n,
@@ -224,6 +218,7 @@ class MenuTabView extends StatelessWidget {
 
                   case ProfileTab.signing:
                     child = card(
+                      key: const Key('Signing'),
                       icon: Icons.lock,
                       title: 'label_login_options'.l10n,
                       subtitle: 'label_login_section_hint'.l10n,
@@ -247,14 +242,14 @@ class MenuTabView extends StatelessWidget {
                     break;
 
                   case ProfileTab.calls:
-                    if (PlatformUtils.isMobile) {
-                      return const SizedBox();
-                    } else {
+                    if (PlatformUtils.isDesktop && PlatformUtils.isWeb) {
                       child = card(
                         icon: Icons.call,
                         title: 'label_calls'.l10n,
                         subtitle: 'label_calls_displaying'.l10n,
                       );
+                    } else {
+                      return const SizedBox();
                     }
                     break;
 
@@ -272,6 +267,7 @@ class MenuTabView extends StatelessWidget {
 
                   case ProfileTab.language:
                     child = card(
+                      key: const Key('Language'),
                       icon: Icons.language,
                       title: 'label_language'.l10n,
                       subtitle: L10n.chosen.value?.name ??
@@ -294,6 +290,7 @@ class MenuTabView extends StatelessWidget {
 
                   case ProfileTab.danger:
                     child = card(
+                      key: const Key('DangerZone'),
                       icon: Icons.dangerous,
                       title: 'label_danger_zone'.l10n,
                       subtitle: 'label_delete_account'.l10n,